Support read snapshot in curve client for the purpose of snapshot consistency check

+++ src/client/libcurve_file.cpp
@@ -193,17 +193,28 @@ int FileClient::Open(const std::string& filename,
                      const UserInfo_t& userinfo,
                      const OpenFlags& openflags) {
     LOG(INFO) << "Opening filename: " << filename << ", flags: " << openflags;
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if(isSnapshot) {
+        LOG(INFO) << "Opening Snapshot sn: " << sn << ", realfilename: " << realfilename;
+    }
+
     FileInstance* fileserv = FileInstance::NewInitedFileInstance(
-        clientconfig_.GetFileServiceOption(), mdsClient_, filename, userinfo,
-        openflags, false);
+        clientconfig_.GetFileServiceOption(), mdsClient_, realfilename, userinfo,
+        openflags, isSnapshot);
     if (fileserv == nullptr) {
         LOG(ERROR) << "NewInitedFileInstance fail";
         return -1;
     }
 
-    int ret = fileserv->Open(filename, userinfo);
+    if(isSnapshot) {
+        fileserv->SetReadSnapshotSn(sn);
+    }
+
+    int ret = fileserv->Open(realfilename, userinfo);
     if (ret != LIBCURVE_ERROR::OK) {
-        LOG(ERROR) << "Open file failed, filename: " << filename
+        LOG(ERROR) << "Open file failed, filename: " << realfilename
                    << ", retCode: " << ret;
         fileserv->UnInitialize();
         delete fileserv;
@@ -215,10 +226,10 @@ int FileClient::Open(const std::string& filename,
     {
         WriteLockGuard lk(rwlock_);
         fileserviceMap_[fd] = fileserv;
-        fileserviceFileNameMap_[filename] = fileserv;
+        fileserviceFileNameMap_[realfilename] = fileserv;
     }
 
-    LOG(INFO) << "Open success, filname = " << filename << ", fd = " << fd;
+    LOG(INFO) << "Open success, filname = " << realfilename << ", fd = " << fd;
     openedFileNum_ << 1;
 
     return fd;
@@ -425,6 +436,13 @@ int FileClient::Rename(const UserInfo_t& userinfo,
 int FileClient::Extend(const std::string& filename,
     const UserInfo_t& userinfo, uint64_t newsize) {
     LIBCURVE_ERROR ret;
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if(isSnapshot) {
+        LOG(INFO) << "Extend Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    }  
     if (mdsClient_ != nullptr) {
         ret = mdsClient_->Extend(filename, userinfo, newsize);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
@@ -441,6 +459,13 @@ int FileClient::Extend(const std::string& filename,
 int FileClient::Unlink(const std::string& filename,
     const UserInfo_t& userinfo, bool deleteforce) {
     LIBCURVE_ERROR ret;
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if(isSnapshot) {
+        LOG(INFO) << "Unlink Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    }  
     if (mdsClient_ != nullptr) {
         ret = mdsClient_->DeleteFile(filename, userinfo, deleteforce);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
@@ -459,10 +484,16 @@ int FileClient::StatFile(const std::string& filename,
     FInfo_t fi;
     FileEpoch_t fEpoch;
     int ret;
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if(isSnapshot) {
+        LOG(INFO) << "StatFile Snapshot sn: " << sn << ", realfilename: " << realfilename;
+    }    
     if (mdsClient_ != nullptr) {
-        ret = mdsClient_->GetFileInfo(filename, userinfo, &fi, &fEpoch);
+        ret = mdsClient_->GetFileInfo(realfilename, userinfo, &fi, &fEpoch);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
-            << "StatFile failed, filename: " << filename << ", ret" << ret;
+            << "StatFile failed, filename: " << realfilename << ", ret" << ret;
     } else {
         LOG(ERROR) << "global mds client not inited!";
         return -LIBCURVE_ERROR::FAILED;
